backend/requirements.txt
Flask
flask-cors
python-dotenv
openai>=1.0.0
gunicorn